Generate outputs from two configs.

+++ convert-arts-to-nc.sh
@@ -2,12 +2,14 @@
 
 [[ ! -d out ]] && mkdir out
 [[ ! -d out ]] && exit 1
+[[ ! -d template ]] && mkdir template
+[[ ! -d template ]] && exit 1
 
 ln -sf in/atmdata/multiple_input4MIPs_radiation_RFMIP_UColorado-RFMIP-1-1_none.nc
 
 ./generate-output-file-templates.py --source_id ARTS-2-3 
 rm -f rs[du]_Efx_ARTS-2-3_rad-irf_r1i1p1f1_gn.nc
-mv rl[du]_Efx_ARTS-2-3_rad-irf_r1i1p1f1_gn.nc out/
+mv rl[du]_Efx_ARTS-2-3_rad-irf_r1i1p1f1_gn.nc template/
 
 python convert2nc.py
